+++ client/src/main/java/replicant/ResponseHandler.java
@@ -1,8 +1,6 @@
 package replicant;
 
-import jsinterop.base.Any;
-
 public interface ResponseHandler
 {
-  void onResponse( Any response );
+  void onResponse( Object response );
 }