+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Remove legacy gwt-rpc infrastructure.
+* Make the response field settable from the create method and make sure it is an Object not an Any for JVM compatibility.
 
 ### [v6.146](https://github.com/replicant4j/replicant/tree/v6.146) (2025-08-12) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.145...v6.146)